fastapi_infer/requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
pillow
